+++ packages/applications/ssr/src/components/pages/réseau/raccordement/transmettre/transmettreDemandeComplèteRaccordement/TransmettreDemandeComplèteRaccordement.form.tsx
@@ -14,7 +14,7 @@ import { ValidationErrors } from '@/utils/formAction';
 import {
   GestionnaireRéseauSelect,
   GestionnaireRéseauSelectProps,
-} from '../../modifier/modifierGestionnaireRéseauRaccordement/GestionnaireRéseauSelect';
+} from '../../GestionnaireRéseauSelect';
 
 import {
   transmettreDemandeComplèteRaccordementAction,
